refactor: split views for sandbox data by type
Split config for room- / thread-based uploads

  'rooms_upload_path' / 'threads_upload_path' IC fields replace
  'upload_path', which now raises (previously implicitly enabled both).

Split `soliplex.views.file_uploads' into:

- 'soliplex.views.room_file_uploads'
- 'soliplex.views.thread_file_uploads'
- 'soliplex.views.sandbox_workdirs'

Closes #1106.
