+++ example/installation-openai.yaml
@@ -353,7 +353,8 @@ title_agent_config_id: "title"
 #==========================================================================
 # Upload directory configuraition
 #==========================================================================
-upload_path: ../uploads
+rooms_upload_path: ../uploads/rooms
+threads_upload_path: ../uploads/threads
 
 #==========================================================================
 # Thread peristence:  SQLAlchemy DBURIs
