+++ example/minimal-openai.yaml
@@ -306,7 +306,8 @@ title_agent_config_id: "title"
 #==========================================================================
 # Upload directory configuraition
 #==========================================================================
-upload_path: ../uploads
+rooms_upload_path: ../uploads/rooms
+threads_upload_path: ../uploads/threads
 
 #==========================================================================
 # Thread peristence:  SQLAlchemy DBURIs
