+++ src/soliplex/config/routing.py
@@ -187,7 +187,9 @@ _DEFAULT_ROUTER_NAMES = {
     "authn": "soliplex.views.authn.router",
     "authz": "soliplex.views.authz.router",
     "completions": "soliplex.views.completions.router",
-    "file_uploads": "soliplex.views.file_uploads.router",
+    "room_file_uploads": "soliplex.views.room_file_uploads.router",
+    "thread_file_uploads": "soliplex.views.thread_file_uploads.router",
+    "sandbox_workdirs": "soliplex.views.sandbox_workdirs.router",
     "installation": "soliplex.views.installation.router",
     "log_ingest": "soliplex.views.log_ingest.router",
     "quizzes": "soliplex.views.quizzes.router",
