+++ tests/unit/views/test_room_file_uploads.py
@@ -1,0 +1,312 @@
+import contextlib
+import io
+from unittest import mock
+
+import fastapi
+import pytest
+
+from soliplex import authz
+from soliplex import installation
+from soliplex import loggers
+from soliplex import models
+from soliplex.config import rooms as config_rooms
+from soliplex.views import room_file_uploads as room_views
+
+USER_NAME = "phreddy"
+EMAIL = "phreddy@example.com"
+
+THE_USER_CLAIMS = {"preferred_username": USER_NAME, "email": EMAIL}
+
+TEST_ROOM_ID = "test-room-id"
+TEST_FILENAME = "test_file.txt"
+TEST_CONTENT = b"DEADBEEF"
+
+URL_PREFIX = "http://test.example.com/api"
+
+no_error = contextlib.nullcontext
+
+
+def raises_httpexc(*, match, code) -> pytest.raises:
+    def _check(exc):
+        return exc.status_code == code
+
+    return pytest.raises(fastapi.HTTPException, match=match, check=_check)
+
+
+@pytest.fixture
+def uploads_path(temp_dir):
+    result = temp_dir / "uploads"
+    result.mkdir()
+    return result
+
+
+@pytest.mark.anyio
+@pytest.mark.parametrize(
+    "w_filenames",
+    [
+        [],
+        ["foo.txt"],
+        [f"file_{i_file:03}.txt" for i_file in range(10)],
+    ],
+)
+@pytest.mark.parametrize(
+    "w_upload_path, w_room_path, expectation",
+    [
+        (True, True, no_error(None)),
+        (True, False, no_error(None)),
+        (
+            False,
+            None,
+            raises_httpexc(code=404, match="Room uploads not configured"),
+        ),
+    ],
+)
+@mock.patch("soliplex.views.agui._check_user_in_room")
+async def test_get_uploads_room_only(
+    cuir,
+    uploads_path,
+    w_upload_path,
+    w_room_path,
+    expectation,
+    w_filenames,
+):
+    room_uploads_path = uploads_path / "rooms"
+    room_path = room_uploads_path / TEST_ROOM_ID
+    # Note: this is the name of the view function, and not the path
+    #       to which it is bound.
+    ROUTE_NAME = "get_uploads_room_filename"
+
+    def download_url(name, room_id, filename):
+        assert name == ROUTE_NAME
+        return f"{URL_PREFIX}/v1/uploads/{room_id}/{filename}"
+
+    exp_filename_urls = {}
+
+    if w_room_path:
+        room_path.mkdir(parents=True)
+        (room_path / "ignore_me").mkdir()
+        for filename in w_filenames:
+            file_path = room_path / filename
+            file_path.write_text(f"filename: {filename}")
+            exp_filename_urls[filename] = download_url(
+                ROUTE_NAME, TEST_ROOM_ID, filename
+            )
+
+    request = mock.create_autospec(fastapi.Request)
+    request.url_for.side_effect = download_url
+
+    room_config = mock.create_autospec(config_rooms.RoomConfig)
+    cuir.return_value = room_config
+
+    the_installation = mock.create_autospec(
+        installation.Installation,
+    )
+
+    if w_upload_path:
+        the_installation.rooms_upload_path = str(room_uploads_path)
+    else:
+        the_installation.rooms_upload_path = None
+
+    the_room_authz = mock.create_autospec(authz.RoomAuthorizationPolicy)
+    the_logger = mock.create_autospec(loggers.LogWrapper)
+
+    with expectation as expected:
+        found = await room_views.get_uploads_room(
+            request=request,
+            room_id=TEST_ROOM_ID,
+            the_installation=the_installation,
+            the_room_authz=the_room_authz,
+            the_user_claims=THE_USER_CLAIMS,
+            the_logger=the_logger,
+        )
+
+    if expected is None:
+        assert isinstance(found, models.RoomUploads)
+        assert found.room_id == TEST_ROOM_ID
+
+        if w_room_path:
+            found_files = {f_up.filename: f_up.url for f_up in found.uploads}
+            assert set(found_files) == set(w_filenames)
+
+            for filename in w_filenames:
+                exp_url = exp_filename_urls[filename]
+                assert str(found_files[filename]) == exp_url
+        else:
+            assert found.uploads == []
+
+    the_logger.debug.assert_called_once_with(
+        loggers.UPLOADS_GET_ROOM,
+    )
+
+
+@pytest.mark.anyio
+@pytest.mark.parametrize(
+    "w_upload_path, w_room_path, w_filename, expectation",
+    [
+        (True, True, True, no_error(None)),
+        (
+            True,
+            True,
+            False,
+            raises_httpexc(code=404, match=".*"),
+        ),
+        (
+            True,
+            False,
+            None,
+            raises_httpexc(code=404, match=".*"),
+        ),
+        (
+            False,
+            None,
+            None,
+            raises_httpexc(code=404, match="Room uploads not configured"),
+        ),
+    ],
+)
+@mock.patch("soliplex.views.agui._check_user_in_room")
+async def test_get_uploads_room_filename(
+    cuir,
+    uploads_path,
+    w_upload_path,
+    w_room_path,
+    w_filename,
+    expectation,
+):
+    room_uploads_path = uploads_path / "rooms"
+    room_path = room_uploads_path / TEST_ROOM_ID
+
+    if w_room_path:
+        room_path.mkdir(parents=True)
+
+        if w_filename:
+            file_path = room_path / TEST_FILENAME
+            file_path.write_text(f"filename: {TEST_FILENAME}")
+
+    room_config = mock.create_autospec(config_rooms.RoomConfig)
+    cuir.return_value = room_config
+
+    the_installation = mock.create_autospec(
+        installation.Installation,
+    )
+
+    if w_upload_path:
+        the_installation.rooms_upload_path = str(room_uploads_path)
+    else:
+        the_installation.rooms_upload_path = None
+
+    the_room_authz = mock.create_autospec(authz.RoomAuthorizationPolicy)
+    the_logger = mock.create_autospec(loggers.LogWrapper)
+
+    with expectation as expected:
+        found = await room_views.get_uploads_room_filename(
+            room_id=TEST_ROOM_ID,
+            filename=TEST_FILENAME,
+            the_installation=the_installation,
+            the_room_authz=the_room_authz,
+            the_user_claims=THE_USER_CLAIMS,
+            the_logger=the_logger,
+        )
+
+    if expected is None:
+        assert found == str(file_path)
+
+    the_logger.debug.assert_called_once_with(
+        loggers.UPLOADS_GET_ROOM_FILE,
+    )
+
+
+@pytest.mark.anyio
+@pytest.mark.parametrize(
+    "w_filename, exp_filename",
+    [
+        (TEST_FILENAME, TEST_FILENAME),
+        ("../../../../etc/passwd", "passwd"),
+    ],
+)
+@pytest.mark.parametrize(
+    "w_upload_path, expectation",
+    [
+        (True, no_error(204)),
+        (False, raises_httpexc(code=404, match="Room uploads not configured")),
+    ],
+)
+@pytest.mark.parametrize("w_admin_access", [False, True])
+@mock.patch("soliplex.views.agui._check_user_in_room")
+async def test_post_uploads_room(
+    cuir,
+    uploads_path,
+    w_admin_access,
+    w_upload_path,
+    expectation,
+    w_filename,
+    exp_filename,
+):
+    room_config = mock.create_autospec(config_rooms.RoomConfig)
+    cuir.return_value = room_config
+    upload_file = fastapi.UploadFile(
+        file=io.BytesIO(TEST_CONTENT),
+        filename=w_filename,
+        headers={"Content-Type": "text/plain"},
+    )
+
+    the_installation = mock.create_autospec(
+        installation.Installation,
+    )
+
+    if w_upload_path:
+        the_installation.rooms_upload_path = str(uploads_path / "rooms")
+    else:
+        the_installation.rooms_upload_path = None
+
+    the_admin_users = mock.create_autospec(authz.AdminUserPolicy)
+    the_admin_users.check_admin_access.return_value = w_admin_access
+    the_room_authz = mock.create_autospec(authz.RoomAuthorizationPolicy)
+    the_logger = mock.create_autospec(loggers.LogWrapper)
+    the_authz_logger = mock.create_autospec(loggers.LogWrapper)
+
+    if not w_admin_access:
+        with pytest.raises(fastapi.HTTPException) as exc:
+            await room_views.post_uploads_room(
+                room_id=TEST_ROOM_ID,
+                upload_file=upload_file,
+                the_installation=the_installation,
+                the_admin_users=the_admin_users,
+                the_room_authz=the_room_authz,
+                the_user_claims=THE_USER_CLAIMS,
+                the_logger=the_logger,
+                the_authz_logger=the_authz_logger,
+            )
+
+        assert exc.value.status_code == 403
+        assert exc.value.detail == loggers.AUTHZ_ADMIN_ACCESS_REQUIRED
+
+        the_authz_logger.error.assert_called_once_with(
+            loggers.AUTHZ_ADMIN_ACCESS_REQUIRED
+        )
+
+    else:
+        with expectation as expected:
+            response = await room_views.post_uploads_room(
+                room_id=TEST_ROOM_ID,
+                upload_file=upload_file,
+                the_installation=the_installation,
+                the_admin_users=the_admin_users,
+                the_room_authz=the_room_authz,
+                the_user_claims=THE_USER_CLAIMS,
+                the_logger=the_logger,
+                the_authz_logger=the_authz_logger,
+            )
+
+        if not isinstance(expected, pytest.ExceptionInfo):
+            assert response.status_code == expected
+            exp_file = uploads_path / "rooms" / TEST_ROOM_ID / exp_filename
+            assert exp_file.read_bytes() == TEST_CONTENT
+
+    the_authz_logger.debug.assert_called_once_with(
+        loggers.UPLOADS_POST_ROOM,
+    )
+
+    the_logger.debug.assert_called_once_with(
+        loggers.UPLOADS_POST_ROOM,
+    )
